Refactor repository queries, extend get responses, and improve error handling
This commit refactors the execution of queries in ProjectRepository to handle userId and uuid in a new format. It also expands the response of several getAll methods across various controllers in the officeframe module by adding a size parameter. Moreover, the error handling is enhanced to manage cases where a document is not found. The EmployeeService has also been updated to generate an employee identifier if it hasn't been specified before. Lastly, CORS configuration in the application.properties has been modified to support credentials.