Commit Message:
Refactor GitHub Workflow and Update Dependencies Version

This commit refactors the GitHub workflow for the Keypractica project. The workflow now includes steps for setting up an SSH Agent, stopping the Keypractica service, updating the project from GitHub, building the project, and restarting the service. It also includes a check to ensure the server is running as expected.

Additionally, this commit updates the version of the 'core' and 'officeframe' dependencies in the projects' pom.xml file, linking it to the project version. Minor tweaks are also made in the Ansible playbook and pom.xml files to remove unnecessary lines.